+++ packages/labs/gen-wrapper-angular/goldens/test-element-a/src/element-without-props.ts
@@ -1,0 +1,12 @@
+import {Component, ElementRef, NgZone} from '@angular/core';
+
+import type {ElementWithoutProps as ElementWithoutPropsElement} from '@lit-internal/test-element-a/element-without-props.js';
+import '@lit-internal/test-element-a/element-without-props.js';
+
+@Component({
+  selector: 'element-without-props',
+  template: '<ng-content></ng-content>',
+})
+export class ElementWithoutProps {
+  constructor(_e: ElementRef<ElementWithoutPropsElement>, _ngZone: NgZone) {}
+}